+++ pkg/oc/cli/admin/node/logs_test.go
@@ -1,4 +1,4 @@
-package logs
+package node
 
 import (
 	"bytes"
